+++ ecosystem/sf-indexer/aptos-sf-indexer/Cargo.toml
@@ -11,38 +11,37 @@ publish = false
 
 [dependencies]
 anyhow = "1.0.57"
+
+aptos-crypto = { path = "../../../crates/aptos-crypto" }
+aptos-logger = { path = "../../../crates/aptos-logger" }
+aptos-metrics-core = { path = "../../../crates/aptos-metrics-core" }
+aptos-protos = { path = "../../../crates/aptos-protos" }
+aptos-rest-client = { path = "../../../crates/aptos-rest-client" }
+aptos-types = { path = "../../../types" }
+async-stream = "0.3"
 async-trait = "0.1.53"
 bigdecimal = { version = "0.1.2", features = ["serde"] }
+bytes = "1.1.0"
 chrono = { version = "0.4.19", default-features = false, features = ["clock", "serde"] }
 clap = "3.1.17"
 diesel = { version = "1.4.8", features = ["chrono", "postgres", "r2d2", "numeric", "serde_json"] }
 diesel_migrations = { version = "1.4.0", features = ["postgres"] }
 field_count = "0.1.1"
+futures = "0.3.21"
+http = "0.2.3"
 once_cell = "1.10.0"
+pbjson = "0.4.0"
+prost = "0.10.4"
+prost-types = "0.10.1"
 reqwest = { version = "0.11.10", features = ["json", "cookies"] }
 reqwest-middleware = { version = "0.1.6" }
 reqwest-retry = { version = "0.1.5" }
 serde = { version = "1.0.137", features = ["derive"] }
 serde_json = "1.0.81"
 tokio = { version = "1.18.2", features = ["full", "time"] }
-url = "2.2.2"
-prost = "0.10.1"
-prost-types = "0.10.1"
-tonic = { version = "0.7.1", features = ["tls-roots"] }
-futures = "0.3.21"
-http = "0.2.3"
-async-stream = "0.3"
 tokio-retry = "0.3.0"
-
-aptos-crypto = { path = "../../../crates/aptos-crypto" }
-aptos-logger = { path = "../../../crates/aptos-logger" }
-aptos-metrics-core = { path = "../../../crates/aptos-metrics-core" }
-aptos-types = { path = "../../../types" }
-aptos-rest-client = { path = "../../../crates/aptos-rest-client" }
-aptos-protos = { path = "../../../crates/aptos-protos" }
-
-[build-dependencies]
-tonic-build = { version = "0.7.2", features = ["prost"] }
+tonic = { version = "0.7.2", features = ["tls-roots", "transport", "prost", "compression", "codegen"] }
+url = "2.2.2"
 
 [[bin]]
 name = "aptos-sf-indexer"
