+++ ecosystem/sf-indexer/aptos-sf-indexer/buf.work.yaml
@@ -1,0 +1,3 @@
+version: v1
+directories:
+  - proto
