+++ ecosystem/sf-indexer/aptos-sf-indexer/build.rs
@@ -2,9 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 fn main() {
-    println!("cargo:rerun-if-changed=proto");
+    /*println!("cargo:rerun-if-changed=proto");
     tonic_build::configure()
         .out_dir("src/proto")
         .compile(&["proto/substreams.proto"], &["proto"])
         .expect("Failed to compile Substreams proto(s)");
+        */
 }
