--- ecosystem/sf-indexer/aptos-sf-indexer/src/proto/google.protobuf.rs
@@ -1,4 +1,0 @@
-
-// Copyright (c) Aptos
-// SPDX-License-Identifier: Apache-2.0
-
